Extract IPC tests from devdax tests
Extract the IPC tests from the devdax tests to a separate file,
because the IPC tests do not work with the proxy library yet
and they should be run separately.
