+++ test/provider_devdax_memory.cpp
@@ -12,17 +12,10 @@
 #include "base.hpp"
 
 #include "cpp_helpers.hpp"
-#include "ipcFixtures.hpp"
 #include "test_helpers.h"
 
 #include <umf/memory_provider.h>
 #include <umf/providers/provider_devdax_memory.h>
-#ifdef UMF_POOL_JEMALLOC_ENABLED
-#include <umf/pools/pool_jemalloc.h>
-#endif
-#ifdef UMF_POOL_SCALABLE_ENABLED
-#include <umf/pools/pool_scalable.h>
-#endif
 
 using umf_test::test;
 
@@ -357,41 +350,3 @@ TEST_F(test, create_wrong_size_0) {
     EXPECT_EQ(ret, UMF_RESULT_ERROR_INVALID_ARGUMENT);
     EXPECT_EQ(hProvider, nullptr);
 }
-
-HostMemoryAccessor hostAccessor;
-
-static std::vector<ipcTestParams> getIpcProxyPoolTestParamsList(void) {
-    std::vector<ipcTestParams> ipcProxyPoolTestParamsList = {};
-
-    char *path = getenv("UMF_TESTS_DEVDAX_PATH");
-    if (path == nullptr || path[0] == 0) {
-        // Test skipped, UMF_TESTS_DEVDAX_PATH is not set
-        return ipcProxyPoolTestParamsList;
-    }
-
-    char *size = getenv("UMF_TESTS_DEVDAX_SIZE");
-    if (size == nullptr || size[0] == 0) {
-        // Test skipped, UMF_TESTS_DEVDAX_PATH is not set
-        return ipcProxyPoolTestParamsList;
-    }
-
-    ipcProxyPoolTestParamsList = {
-        {umfProxyPoolOps(), nullptr, umfDevDaxMemoryProviderOps(),
-         &defaultDevDaxParams, &hostAccessor, true},
-#ifdef UMF_POOL_JEMALLOC_ENABLED
-        {umfJemallocPoolOps(), nullptr, umfDevDaxMemoryProviderOps(),
-         &defaultDevDaxParams, &hostAccessor, false},
-#endif
-#ifdef UMF_POOL_SCALABLE_ENABLED
-        {umfScalablePoolOps(), nullptr, umfDevDaxMemoryProviderOps(),
-         &defaultDevDaxParams, &hostAccessor, false},
-#endif
-    };
-
-    return ipcProxyPoolTestParamsList;
-}
-
-GTEST_ALLOW_UNINSTANTIATED_PARAMETERIZED_TEST(umfIpcTest);
-
-INSTANTIATE_TEST_SUITE_P(DevDaxProviderDifferentPoolsTest, umfIpcTest,
-                         ::testing::ValuesIn(getIpcProxyPoolTestParamsList()));
